[update] organize build structure and format code
